Move non-fastboot head.js back into browser/head.js to fix error created when the project gets built
