+++ app-lt-2-10/instance-initializers/browser/head.js
@@ -1,12 +1,12 @@
-import ENV from '../config/environment';
+import ENV from '../../config/environment';
 
-export function initialize(owner) {
+export function initialize(instance) {
   if (ENV['ember-cli-head'] && ENV['ember-cli-head']['suppressBrowserRender']) { return true; }
 
   // clear fast booted head (if any)
   let startMeta = document.querySelector('meta[name="ember-cli-head-start"]')
   let endMeta = document.querySelector('meta[name="ember-cli-head-end"]')
-  if (startMeta && endMeta) {
+  if (startMeta && startMeta) {
     let el = startMeta.nextSibling
     while(el && el !== endMeta) {
       document.head.removeChild(el);
@@ -16,7 +16,9 @@ export function initialize(owner) {
     document.head.removeChild(endMeta);
   }
 
-  const component = owner.lookup('component:head-layout');
+  const container = instance.lookup ? instance : instance.container;
+  // const renderer = container.lookup('renderer:-dom');
+  const component = container.lookup('component:head-layout');
   component.appendTo(document.head);
 }
